requirements.txt
numpy
torch
torchvision
matplotlib
tqdm
pyyaml
tensorboard
